+++ core/common/core_types.h
@@ -1,0 +1,14 @@
+#ifndef CORE_TYPES_H
+#define CORE_TYPES_H
+
+typedef struct {
+    float x;
+    float y;
+} CORE_POINT2F;
+
+typedef struct {
+    float yaw;
+    float pitch;
+} CORE_ATTITUDE2F;
+
+#endif /* CORE_TYPES_H */